General mapping function. Specify a map type and/or a route id.

General mapping function. Specify a map type and/or a route id.

map_gtfs(gtfs_obj, route_ids = NULL, service_ids = NULL, shape_ids = NULL,
  agency_name = NULL, include_stops = TRUE, only_stops = FALSE,
  stop_details = FALSE, stop_opacity = 0.5, route_opacity = 0.75,
  route_colors = NULL)

Arguments

gtfs_obj
A GTFS list object with components agency_df, etc.

route_ids
Vector (Character). IDs for routes of interest.

service_ids
Vector (Character). Service IDs. NULL by Default.

shape_ids
Vector (Character). Shape IDs. NULL by Default.

agency_name
Character. Provide the name of the agency whose routes are being mapped.

include_stops
Boolean. Whether to layer on stops to the route shape. Default is TRUE.

only_stops
Boolean. Whether to map only stops, no routes. Overrides `include_stops`. Default is FALSE.

stop_details
Boolean. Whether to generate detail stop information. Default is FALSE.

stop_opacity
Numeric. Value must be between 0 and 1. Defaults is 0.5.

route_opacity
Numeric. Value must be between 0 and 1. Default is .75.

route_colors
Character. Names of colors (e.g. "blue") or hex values (e.g. '#000000'). Default is NULL.

Value

Leaflet map object with all stop lat/long values plotted for a route.